organize code into separate modules
