+++ src/datagrunt/__init__.py
@@ -28,7 +28,8 @@ __author__ = "Martin Graham"
 __license__ = "MIT"
 
 # Import key classes, functions, or submodules that should be available at the package level
-from .csvfile import CSVReader, CSVWriter
+from .csvfiles import CSVReader, CSVWriter
+
 
 # You can define __all__ to specify what gets imported with "from package import *"
 __all__ = ['CSVReader', 'CSVWriter']
